+++ openbank-admin-ui/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.133.0](https://github.com/JiRaska/open-bank-oss/compare/admin-ui-v0.132.0...admin-ui-v0.133.0) (2026-08-14)
+
+
+### Features
+
+* **admin-ui:** unify operator surfaces ([#4842](https://github.com/JiRaska/open-bank-oss/issues/4842)) ([519f1e4](https://github.com/JiRaska/open-bank-oss/commit/519f1e406201fdbe7d5df999faa8da91423bfdcb))
+
+## [0.132.0](https://github.com/JiRaska/open-bank-oss/compare/admin-ui-v0.131.0...admin-ui-v0.132.0) (2026-08-14)
+
+
+### Features
+
+* **admin-ui:** unify operator workspace ([#4821](https://github.com/JiRaska/open-bank-oss/issues/4821)) ([c2415dd](https://github.com/JiRaska/open-bank-oss/commit/c2415ddea8f4bd4d82ceb57c86fb8a04b6386548))
+
 ## [0.131.0](https://github.com/JiRaska/open-bank-oss/compare/admin-ui-v0.130.0...admin-ui-v0.131.0) (2026-08-14)